requirements.txt
# requirements.txt
flask[async]
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-marshmallow
marshmallow-sqlalchemy
asyncpg
python-dotenv